add unittest

+++ src/app/controllers/ProjectController.js
@@ -4,7 +4,9 @@ class ProjectController {
     async create(req, res) {
         try {
             const { name, description, location, user_id, start_date, end_date, quantity, uni_ids } = req.body;
-            console.log(req.body);
+            if (!name) {
+                return res.status(400).json('Invalid Project')
+            }
 
             const query = `INSERT INTO project (name, description, location, user_id, start_date, end_date, quantity) 
                     VALUES ($1, $2, $3, $4, $5, $6, $7)
@@ -25,13 +27,13 @@ class ProjectController {
             uni_ids.map(async (uni_id) => {
                 try {
                     const createProjectUni = await pool.query(projectUniQuery, [response.rows[0].id, uni_id, false]);
-                    console.log('ek');
                 } catch (error) {
                     console.log(error);
+                    return res.status(500).json('Internal Server Error');
                 }
             });
 
-            return res.status(200).json({
+            return res.status(201).json({
                 message: 'Project created successfully',
                 data: {
                     name,
@@ -41,20 +43,24 @@ class ProjectController {
                     start_date,
                     end_date,
                     quantity,
+                    id: response.rows[0].id
                 },
             });
         } catch (error) {
             console.log(error);
+            return res.status(500).json('Internal Server Error');
         }
     }
 
     async getAll(req, res) {
         try {
             const query = 'SELECT * FROM project';
+            console.log('alo')
             const data = await pool.query(query);
             return res.status(200).json(data.rows);
         } catch (error) {
             console.log(error);
+            return res.status(500).json('Internal Server Error');
         }
     }
 
@@ -67,7 +73,7 @@ class ProjectController {
                     message: 'Found project successfully',
                     code: 200,
                     body: {
-                        project: response.rows[0],
+                        project: response,
                     },
                 });
             } else {
@@ -78,6 +84,8 @@ class ProjectController {
             }
         } catch (error) {
             console.log(error);
+            return res.status(500).json('Internal Server Error');
+
         }
     }
 
@@ -98,6 +106,7 @@ class ProjectController {
             });
         } catch (err) {
             console.log(err);
+            return res.status(500).json('Internal Server Error');
         }
     }
 
@@ -105,26 +114,39 @@ class ProjectController {
         try {
             const id = parseInt(req.params.slug);
             const { isChecked } = req.body;
+  
+
             const query = `
-                UPDATE project SET isChecked = $1 WHERE id = $2
+                UPDATE project SET is_checked = $1 WHERE id = $2
             `;
+            
             const values = [isChecked, id];
             const response = await pool.query(query, values);
 
+            const resProject = (await pool.query('SELECT * from project WHERE id = $1', [id]))
+            if (resProject.rows.length === 0) {
+                return res.status(400).json({message: 'Project not found!'});
+            }       
+
             return res.status(200).json({
                 message: 'Update Project successfully!',
-                data: response,
+                data: resProject.rows[0],
             });
         } catch (err) {
-            console.log(err);
+            return res.status(500).json('Internal Server Error');
         }
     }
 
     async delete(req, res) {
         try {
             const id = parseInt(req.params.slug);
-            const { project_id } = parseInt(req.body);
-            console.log(id);
+
+            const resProject = await pool.query('SELECT * from project where id = $1', [id])
+
+            if (resProject.rows.length === 0) {
+                return res.status(400).json({message: 'Project not found'})
+            }
+
             const deleteProjectQuery = `
                 DELETE FROM project
                 WHERE id = $1
@@ -132,9 +154,14 @@ class ProjectController {
             const response = await pool.query(deleteProjectQuery, [id]);
             return res.status(200).json({
                 message: 'Delete Project successfully!',
-                data: response,
+                data: {
+                    project_id: id
+                },
             });
-        } catch (error) {}
+        } catch (error) {
+            console.log(error);
+            return res.status(500).json('Internal Server Error');
+        }
     }
 }
